pygenum: When getting a member instance, use correct key for the lookup
We were using the value from the introspection info as the key directly,
but GEnum uses an int for the value, and the introspection value is a long.

So, values greather than 0x7fffffff were looked up incorrectly.

Fixes #588.

+++ gi/pygenum.c
@@ -170,7 +170,7 @@ pyg_enum_new(PyTypeObject *type, PyObject *args, PyObject *kwargs)
 
     g_type_class_unref(eclass);
 
-    intvalue = PyLong_FromLong(value);
+    intvalue = PyLong_FromLong((int)value);
     ret = PyDict_GetItem(values, intvalue);
     Py_DECREF(intvalue);
     Py_DECREF(values);
